library_management_api/library/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.7
django-crispy-forms==2.3
django-filter==24.3
django-heroku==0.3.1
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
